fix: Infer expression types before renaming identifiers
- Reorder transpile_typed passes so Call callees resolve through TypeCtx::funcs before rename collapses Ident to Raw
- Restore invariant already documented in infer_types Raw arm comment
- 159 unchanged forge_conformance byte-goldens prove no other expression is perturbed

+++ tests/forge/expected/CrossfileValidatorTransform.kt
@@ -13,7 +13,7 @@ class CrossfileValidatorTransform {
     fun validate(rawTemp: UShort): ValidationResult {
         if (rawTemp.toInt() < 0 || rawTemp.toInt() > 4095)
             return ValidationResult(false, "rawTemp_out_of_range")
-        if (!(computeTemperature(rawTemp) > -40 && computeTemperature(rawTemp) < 200))
+        if (!(computeTemperature(rawTemp) > (-40).toDouble() && computeTemperature(rawTemp) < 200.0))
             return ValidationResult(false, "plausibility_failed")
         return ValidationResult(true, "")
     }
